+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v1.0.2
+### Added
+- Added string Extensions for RemoveLeadingSlash and RemoveTrailingSlash
+- Added string Extensions for EnsureLeadingSlash and EnsureTrailingSlash
+- Enumeration for Separators
+
 ## v1.0.1
 ### Fixed
 - Fixed XML Documentation